feat: change type defination
