+++ packages/utils/src/typescript/types.ts
@@ -1,40 +1,17 @@
-import { isArray, isObject, isString } from '@vue/shared';
-import { isNil } from 'lodash-es';
-
-export {
-  isArray,
-  isFunction,
-  isObject,
-  isString,
-  isDate,
-  isPromise,
-  isSymbol,
-} from '@vue/shared';
-export { isVNode } from 'vue';
-
-export const isUndefined = (val: any): val is undefined => val === undefined;
-export const isBoolean = (val: any): val is boolean => typeof val === 'boolean';
-export const isNumber = (val: any): val is number => typeof val === 'number';
-
-export const isEmpty = (val: unknown) =>
-  (!val && val !== 0) ||
-  (isArray(val) && val.length === 0) ||
-  (isObject(val) && !Object.keys(val).length);
-
-export const isElement = (e: unknown): e is Element => {
-  if (typeof Element === 'undefined') return false;
-  return e instanceof Element;
-};
-
-export const isPropAbsent = (prop: unknown): prop is null | undefined => {
-  return isNil(prop);
-};
-
-export const isStringNumber = (val: string): boolean => {
-  if (!isString(val)) {
-    return false;
-  }
-  return !Number.isNaN(Number(val));
-};
-
 export type Timer = ReturnType<typeof setTimeout>;
+
+export type Mutable<T> = {
+  -readonly [Key in keyof T]: T[Key];
+};
+
+export type HTMLElementCustomized<T> = HTMLElement & T;
+
+/**
+ * @deprecated stop to use null
+ * @see {@link https://github.com/sindresorhus/meta/discussions/7}
+ */
+export type Nullable<T> = T | null;
+
+export type Arrayable<T> = T | T[];
+export type Awaitable<T> = Promise<T> | T;
+export type UnknowFunction = (...args: unknown[]) => unknown;
